refactor(ui): split assistant message parts and command panel
Extract assistant timeline helpers and analysis sheet from assistant-message-parts.
Move CommandPanel to command-panel.tsx with settings-nav constants.
Remove dead commented JSX in ChainOfThoughtStep.
Defer setState in chat effects and route session key for ESLint compliance.
