+++ apps/tanstack-start/src/components/chat/assistant-message-helpers.ts
@@ -1,0 +1,64 @@
+import {
+  BrainIcon,
+  CheckIcon,
+  FlaskConicalIcon,
+  GlobeIcon,
+  SearchIcon,
+  WrenchIcon,
+} from "lucide-react";
+
+import type { AssistantTimelineStep } from "./assistant-message-timeline";
+
+export function getAssistantStepIcon(step: AssistantTimelineStep) {
+  if (step.kind === "reasoning") {
+    return BrainIcon;
+  }
+
+  if (step.kind === "source") {
+    return GlobeIcon;
+  }
+
+  if (step.label.toLowerCase().includes("search")) {
+    return SearchIcon;
+  }
+
+  if (step.toolName?.toLowerCase() === "analysis_workspace") {
+    return FlaskConicalIcon;
+  }
+
+  return WrenchIcon;
+}
+
+export function getChainOfThoughtHeaderState(steps: AssistantTimelineStep[]) {
+  const activeStep =
+    [...steps].reverse().find((step) => step.status === "active") ??
+    [...steps].reverse().find((step) => step.status === "pending");
+
+  if (activeStep) {
+    return {
+      icon: getAssistantStepIcon(activeStep),
+      label: activeStep.summary ?? activeStep.label,
+      status: activeStep.status,
+    } as const;
+  }
+
+  const errorCount = steps.filter((step) => step.status === "error").length;
+  const completeCount = steps.filter(
+    (step) => step.status === "complete",
+  ).length;
+  const totalCount = steps.length;
+
+  if (errorCount > 0) {
+    return {
+      icon: WrenchIcon,
+      label: `Completed ${completeCount} of ${totalCount} steps`,
+      status: "error",
+    } as const;
+  }
+
+  return {
+    icon: CheckIcon,
+    label: `Completed ${totalCount} step${totalCount === 1 ? "" : "s"}`,
+    status: "complete",
+  } as const;
+}
